Updated for lucene 9.2.0

+++ pso-flint-lucene/src/test/java/org/pageseeder/flint/lucene/search/AutoSuggestTest.java
@@ -34,14 +34,14 @@ public class AutoSuggestTest {
 
   private static LuceneLocalIndex index;
   private static IndexManager manager;
-  
+
   @BeforeClass
   public static void init() {
     // clean up previous test's data
-    for (File f : indexRoot.listFiles()) f.delete();
+    if (indexRoot.exists()) for (File f : indexRoot.listFiles()) f.delete();
     indexRoot.delete();
     try {
-      index = new LuceneLocalIndex(indexRoot, new StandardAnalyzer(), documents);
+      index = new LuceneLocalIndex(indexRoot, "autosuggest", new StandardAnalyzer(), documents);
       index.setTemplate("xml", template.toURI());
     } catch (Exception ex) {
       ex.printStackTrace();
