+++ pso-flint-lucene/src/test/java/org/pageseeder/flint/lucene/search/FacetsTest.java
@@ -29,14 +29,14 @@ public class FacetsTest {
 
   private static LuceneLocalIndex index;
   private static LocalIndexManager manager;
-  
+
   @BeforeClass
   public static void init() {
     // clean up previous test's data
-    for (File f : indexRoot.listFiles()) f.delete();
+    if (indexRoot.exists()) for (File f : indexRoot.listFiles()) f.delete();
     indexRoot.delete();
     try {
-      index = new LuceneLocalIndex(indexRoot, new StandardAnalyzer(), documents);
+      index = new LuceneLocalIndex(indexRoot, "facets", new StandardAnalyzer(), documents);
       index.setTemplate("xml", template.toURI());
     } catch (Exception ex) {
       ex.printStackTrace();
